service/cicloguia/requirements.txt
fastapi==0.70.1
pydantic==1.8.2
icecream==2.1.1
pytest==6.2.5
mock~=4.0.3
setuptools==60.1.0
google-auth
google-cloud-firestore
uvicorn[standard]
boto3